+++ agb/src/display/tiled/vram_manager.rs
@@ -461,4 +461,18 @@ impl VRamManager {
             self.set_background_palette(palette_index as u8, entry);
         }
     }
+
+    /// Gets the index of the colour for a given background palette, or None if it doesn't exist
+    #[must_use]
+    pub fn find_colour_index_16(&self, palette_index: usize, colour: u16) -> Option<usize> {
+        assert!(palette_index < 16);
+
+        (0..16).find(|i| PALETTE_BACKGROUND.get(palette_index * 16 + i) == colour)
+    }
+
+    /// Gets the index of the colour in the entire background palette, or None if it doesn't exist
+    #[must_use]
+    pub fn find_colour_index_256(&self, colour: u16) -> Option<usize> {
+        (0..256).find(|&i| PALETTE_BACKGROUND.get(i) == colour)
+    }
 }